# Turn off optimization mode when compiling tests for K/N