--- a/kotlinx-coroutines-test/build.gradle.kts
+++ b/kotlinx-coroutines-test/build.gradle.kts
@@ -2,12 +2,6 @@ import org.jetbrains.kotlin.gradle.plugin.mpp.*
 import org.jetbrains.kotlin.gradle.targets.js.dsl.*
 
 kotlin {
-    targets.withType(KotlinNativeTargetWithTests::class.java).configureEach {
-        binaries.getTest("DEBUG").apply {
-            optimized = true
-        }
-    }
-
     sourceSets {
         jvmTest {
             dependencies {