Optimization was turned on due to performance problems. However, with new K/N memory manager these performance problems seem to be fixed. On the other hand, compiling in optimized mode significantly increases the compiler's memory consumption, so it's better to turn it off.
In particular, we (in K/N team) are having some problems with updating to LLVM 16 due to increased memory consumption (it's still being investigated what causes it), more details here: https://youtrack.jetbrains.com/issue/KT-68418/Native-updating-to-LLVM-16-makes-kotlinx.coroutines-build-in-Aggregate-unstable. This commit workarounds it.