@homuroll 2024-06-25
Some numbers:
- 2nd run of `./gradlew clean :kotlinx-coroutines-core:linkDebugTestMacosArm64` sped up from ~69 sec to ~19 sec
- but the running time of the test binary increased from ~38 sec to ~49.5 sec
The difference between the optimized binary and the non-optimized one is noticeable but I believe it's not that bad as it's used to be and won't be a problem.